github actions: add pypi upload support

+++ upload.sh
@@ -12,8 +12,14 @@ if [ "$@" == '--local' ]; then
   python3 -m twine upload --repository $PYPI --repository-url $TWINE_REPOSITORY \
     -u $TWINE_USERNAME -p $TWINE_PASSWORD --non-interactive --skip-existing dist/*
   python3 setup.py clean --all
-else
+elif [ "$@" == '--docker' ]; then
   set -x
-  ./build.sh
+  /bin/bash docker/build.sh
   docker run --rm -it --env-file env.sh automon "$@"
+elif [ "$@" == '--github' ]; then
+  set -x
+  python3 setup.py sdist bdist_wheel
+  twine check dist/*
+  python3 -m twine upload --repository $PYPI --repository-url $TWINE_REPOSITORY \
+    -u $TWINE_USERNAME -p $TWINE_PASSWORD --non-interactive --skip-existing dist/*
 fi
